feat: add impl for cell
