+++ src/biz/collab/mod.rs
@@ -1,3 +1,4 @@
 pub mod folder_view;
 pub mod ops;
 pub mod publish_outline;
+pub mod utils;
